+++ bolt-modules/boltlib/spec/functions/run_command_spec.rb
@@ -14,6 +14,7 @@ describe 'run_command' do
   around(:each) do |example|
     Puppet[:tasks] = tasks_enabled
     Puppet.override(bolt_executor: executor, bolt_inventory: inventory) do
+      inventory.stubs(:version).returns(1)
       example.run
     end
   end