+++ bolt-modules/boltlib/spec/functions/run_task_spec.rb
@@ -35,6 +35,7 @@ describe 'run_task' do
     executor.stubs(:noop).returns(false)
 
     Puppet.override(bolt_executor: executor, bolt_inventory: inventory) do
+      inventory.stubs(:version).returns(1)
       example.run
     end
   end